chore: clippy warnings and misc changes
There were a couple of new warnings from Clippy that needed fixed.

The first was a suggestion that the `proposals_sig_share` field was boxed because there was a large
size difference between the `SuperMajority` and `Merge` variants, and the second was to use the
`write!` macro rather than `format!` because it involved extra allocation on the `String` type.

This commit also has a few misc changes:

* Switch off verbose output for cargo in the PR workflow because it makes the build output noisy.
* Remove superfluous comments from the PR workflow file.
* Use a .msc file extension for a piece of test data. I think this was omitted by mistake and it
  means the file doesn't get ignored as a *.msc, which is in the .gitignore.